Added URLSessionTaskPublisherDelegate
